% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon.R
\name{iota}
\alias{iota}
\title{Prefix function binding for \code{:}}
\description{
\code{iota(a, b)} is equivalent to \code{a : b} for creating regular sequences of numbers.
\code{iota} does not support the use of \code{:} for interaction terms in regression formulae.
}
