% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{add}
\alias{add}
\alias{minus}
\alias{mult}
\alias{div}
\alias{pow}
\alias{mod}
\alias{floordiv}
\alias{dot}
\alias{gt}
\alias{geq}
\alias{lt}
\alias{leq}
\alias{eq}
\alias{neq}
\alias{not}
\alias{and}
\alias{or}
\alias{xor}
\alias{item_and}
\alias{item_or}
\alias{iota}
\alias{index}
\alias{item}
\alias{isin}
\alias{union}
\alias{intersect}
\alias{setdiff}
\title{Prefix functions}
\usage{
add(1, 2)

minus(10, 5)

mult(a, b = NULL)

div(a, b = NULL)

pow(a, b = NULL)

mod(a, b = NULL)

floordiv(a, b = NULL)

dot(a, b = NULL)

gt(a, b = NULL)

geq(a, b = NULL)

lt(a, b = NULL)

leq(a, b = NULL)

eq(a, b = NULL)

neq(a, b = NULL)

not(a, b = NULL)

and(a, b = NULL)

or(a, b = NULL)

xor(a, b = NULL)

item_and(a, b = NULL)

item_or(a, b = NULL)

iota(a, b = NULL)

index()

item()

isin(a, b = NULL)

union(a, b = NULL)

intersect(a, b = NULL)

setdiff(a, b = NULL)
}
\description{
Prefix functions for built-in binary operators.

Example, write \code{1 + 2} as add(1, 2).

Functions for binary operations can be partially applied if only one argument is provided. For example,

add_one = add(1)
}
